+++ CHANGELOG.md
@@ -1,2 +1,14 @@
 ## [0.0.1]
-- Initial release, 10.04.2020
+- Initial release, 10.04.2020.
+
+## [0.0.2]
+- Trying to figure out why Flutter Pub website doesn't display gifs in the README.md, 10.04.2020.
+
+## [0.1.0]
+- Fix gifs in the README.md, 11.04.2020.
+
+## [0.1.1]
+- add MultiCircularMenu widget, 2.05.2020.
+
+## [0.1.2]
+- remove CircularMenuItem default padding, 21.05.2020.